backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
requests==2.31.0
python-dateutil==2.8.2
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
Pillow==10.2.0 